Moved some classes from Database to Database\Runtime
