+++ Gishiki/Database/Adapters/Sqlite.php
@@ -19,8 +19,8 @@ namespace Gishiki\Database\Adapters;
 
 use Gishiki\Database\RelationalDatabaseInterface;
 use Gishiki\Database\DatabaseException;
-use Gishiki\Database\SelectionCriteria;
-use Gishiki\Database\ResultModifier;
+use Gishiki\Database\Runtime\SelectionCriteria;
+use Gishiki\Database\Runtime\ResultModifier;
 use Gishiki\Algorithms\Collections\GenericCollection;
 use Gishiki\Database\Adapters\Utils\SQLBuilder;
