+++ Gishiki/Database/Runtime/FieldOrdering.php
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 *****************************************************************************/
 
-namespace Gishiki\Database;
+namespace Gishiki\Database\Runtime;
 
 /**
  * A collection of orders that can be applied to a field.
